- 설정파일 수정 - 불필요한 소스 제거

+++ src/main/kotlin/com/jeonguk/web/domain/Order.kt
@@ -1,7 +1,9 @@
 package com.jeonguk.web.domain
 
+import java.math.BigInteger
+
 data class Order(
-        var orderId: Long? = null,
+        var orderId: BigInteger? = null,
         var userId: Int? = null,
         var status: String? = null
 )
